requirements.txt
requests
plotly
numpy
skyfield
python-dotenv